requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
airtable-python-wrapper==0.15.3
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
assemblyai==0.31.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
config==0.5.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-anthropic==0.1.20
langchain-community==0.2.9
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langgraph==0.1.9
langsmith==0.1.93
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
notion-client==2.2.1
numpy==1.26.4
openai==1.36.0
orjson==3.10.6
packaging==24.1
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
pyairtable==2.3.3
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
python-telegram-bot==21.4
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tavily-python==0.3.5
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.4